pip-requirements.txt
connexion[swagger-ui]
Flask-SQLAlchemy
ipython
mysqlclient  // maybe unused
mysql-connector-python
sqlalchemy
sqlalchemy-utils
